Time: 20 ms (68.2%), Space: 70.6 MB (12.63%) - LeetHub

+++ 1482-minimum-number-of-days-to-make-m-bouquets/1482-minimum-number-of-days-to-make-m-bouquets.cpp
@@ -16,7 +16,7 @@ public:
         return totalBkt;
     }
     int minDays(vector<int>& arr, int m, int k) {
-        long long n=arr.size();
+        int n=arr.size();
         int st=*min_element(arr.begin(),arr.end());
         int end=*max_element(arr.begin(),arr.end());
          if((double)m*k >(double)n)return -1;
